Refactored non-generated dto classes to their own file to keep the auto-generated dto file as-is. Added a lazy NSAllowsArbitraryLoads into Info.plist due to http requests getting blocked (a new iOS 9 requirement), and without knowing in advance all the various domains/sub-domains, the flag opens up the http protocol without restriction. The other route (if all requests are known) would be to add all these domains/sub-domains into the NSExceptionDomains dictionary value list. Cleaned up a couple lines of code in the AppData class.
