+++ src/TechStacks/Info.plist
@@ -4,6 +4,8 @@
 <dict>
 	<key>NSAppTransportSecurity</key>
 	<dict>
+		<key>NSAllowsArbitraryLoads</key>
+		<true/>
 		<key>NSExceptionDomains</key>
 		<dict>
 			<key>techstacks.io</key>
@@ -14,8 +16,6 @@
 				<false/>
 			</dict>
 		</dict>
-		<key>New item</key>
-		<string></string>
 	</dict>
 	<key>CFBundleDevelopmentRegion</key>
 	<string>en</string>
